Renaming protocol to matrixprotocol
