+++ src/jaxfun/la/__init__.py
@@ -1,8 +1,7 @@
 from .diamatrix import (
     DiaMatrix as DiaMatrix,
-    LUFactors as LUFactors,
     diags as diags,
     diakron as diakron,
 )
 from .matrix import Matrix as Matrix
-from .protocol import MatrixProtocol as MatrixProtocol
+from .matrixprotocol import MatrixProtocol as MatrixProtocol
