+++ src/jaxfun/la/matrixprotocol.py
